+++ .env.example
@@ -1,3 +1,2 @@
 REDIS_URL=redis://127.0.0.1:6379
 DEFAULT_QUEUE=compiler
-EMSCRIPTEN_BINARY=/emsdk/emscripten/1.38.20/emcc
