+++ README.md
@@ -2,15 +2,15 @@
 A playground for compiling and running WebAssembly applications.
 
 ## Getting Started
-1. git clone ```https://github.com/wasm-lab/wasm-playground.git```
-2. cd ```wasm-playground```
+1. git clone ```https://github.com/wasm-lab/wasm-playground-server.git```
+2. cd ```wasm-playground-server```
 3. npm install
 
 To run various commands, use the following commands:
-* start - Builds the project and runs the server
-* build:dev - Builds both the server and the client projects
-* build:client:dev - Builds the client project on webpack.dev.js config
-* build:server:dev - Builds the client project on webpack.server.js config
+* start - Runs the production build and starts the server
+* start:dev - Runs the development build starts the server
+* build:dev - Builds the server using webpack.dev.js
+* build:prod - Builds the server using webpack.prod.js
 * start:worker - Transpiles and runs the queue worker
 * test - Run the tests
