+++ package.json
@@ -1,15 +1,15 @@
 {
-  "name": "wasm-playground",
+  "name": "wasm-playground-server",
   "version": "0.1.0",
   "description": "A playground for compiling and running WebAssembly applications",
-  "main": "server/server.js",
+  "main": "src/server.js",
   "scripts": {
-    "test": "mocha ./server/tests/*.test.js --require @babel/register --require babel-polyfill",
-    "build:dev": "webpack --mode development --config webpack.server.js && webpack --config webpack.dev.js",
-    "build:client:dev": "webpack --mode development --config webpack.dev.js",
-    "build:server:dev": "webpack --mode development --config webpack.server.js",
-    "start:worker": "nodemon server/worker.js --exec babel-node server/worker.js",
-    "start": "npm run build:dev && node dist/server.js"
+    "test": "mocha ./src/tests/*.test.js --require @babel/register --require babel-polyfill",
+    "build:dev": "webpack --mode development --config webpack.dev.js",
+    "build:prod": "webpack --mode production --config webpack.prod.js",
+    "start:worker": "nodemon src/worker.js --exec babel-node src/worker.js",
+    "start:dev": "npm run build:dev && nodemon dist/server.js",
+    "start": "npm run build:prod && nodemon dist/server.js"
   },
   "keywords": [
     "webassembly",
@@ -29,35 +29,21 @@
     "babel-register": "^6.26.0",
     "chai": "^4.2.0",
     "chai-http": "^4.2.0",
+    "clean-webpack-plugin": "^1.0.0",
+    "mocha": "^5.2.0",
     "nodemon": "^1.18.7",
-    "vue-loader": "^15.4.2",
-    "vue-template-compiler": "^2.5.17",
     "webpack": "^4.26.0",
     "webpack-cli": "^3.1.2",
-    "webpack-dev-middleware": "^3.4.0",
-    "webpack-dev-server": "^3.1.10",
-    "webpack-hot-middleware": "^2.24.3",
     "webpack-merge": "^4.1.4",
     "webpack-node-externals": "^1.7.2"
   },
   "dependencies": {
-    "axios": "^0.18.0",
-    "body-parser": "^1.18.3",
     "bull": "^3.5.2",
-    "clean-webpack-plugin": "^1.0.0",
     "cors": "^2.8.5",
-    "css-loader": "^1.0.1",
     "dotenv": "^6.1.0",
     "express": "^4.16.4",
     "express-zip": "^2.0.1",
     "file-loader": "^2.0.0",
-    "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^3.2.0",
-    "multer": "^1.4.1",
-    "sass-loader": "^7.1.0",
-    "socket.io": "^2.2.0",
-    "socket.io-client": "^2.2.0",
-    "style-loader": "^0.23.1",
-    "vue": "^2.5.17"
+    "multer": "^1.4.1"
   }
 }
