+++ src/controllers/compile.js
@@ -13,14 +13,19 @@ class CompileController {
      * @return {json}
      */
     static compileFile(req, res) {
-        //Run job
+        // Run job
         let inputFile = path.join(__dirname, '../server/uploads/') + req.file.filename;
         let outputPath = path.join(__dirname, '../client/compiled/');
         let outputFileName = req.file.filename.split('.')[0];
 
+        // Create output directory if not exist
+        if (! fs.existsSync(outputPath)){
+            fs.mkdirSync(outputPath);
+        }
+
         let options = ` -o ${outputPath}${outputFileName}.js`;
 
-        let cmd = os.homedir() + process.env.EMSCRIPTEN_BINARY + ` ${inputFile}` + options;
+        const cmd = `emcc ${inputFile} ${options}`;
 
         const child = exec(cmd, (error, stdout, stderr) => {
             if (error) {
