+++ webpack.dev.js
@@ -1,7 +1,7 @@
 const merge = require('webpack-merge');
 const common = require('./webpack.common');
 
-module.exports = merge(common, {
-   mode: 'development',
-   devtool: 'inline-source-map',
+module.exports = merge(common,{
+    mode: 'development',
+    devtool: 'inline-source-map',
 });
