+++ webpack.prod.js
@@ -1,0 +1,7 @@
+const merge = require('webpack-merge');
+const common = require('./webpack.common');
+
+module.exports = merge(common,{
+    mode: 'production',
+    devtool: 'source-map',
+});
