--- webpack.server.js
@@ -1,35 +1,0 @@
-const path = require('path');
-const nodeExternals = require('webpack-node-externals');
-const CleanWebpackPlugin = require('clean-webpack-plugin');
-
-module.exports = {
-    entry: {
-        server: './server/server.js'
-    },
-    output: {
-        filename: '[name].js',
-        path: path.join(__dirname, 'dist'),
-        publicPath: '/'
-    },
-    mode: 'development',
-    target: 'node',
-    node: {
-        __dirname: false,
-        __filename: false,
-    },
-    externals: [nodeExternals()],
-    module: {
-        rules: [
-            {
-                test: /\.js$/,
-                exclude: /node_modules/,
-                use: {
-                    loader: 'babel-loader'
-                }
-            }
-        ]
-    },
-    plugins: [
-        new CleanWebpackPlugin(['dist']),
-    ]
-}
